requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
EbookLib==0.17.1
Flask==2.0.2
Flask-Cors==3.0.10
greenlet==1.1.2
idna==3.3
importlib-metadata==4.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.28
typing_extensions==4.0.1
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
zipp==3.6.0
